Moved all user created UIViewController extensions into UIViewController.swift
