+++ AlphaWallet/Wallet/ViewControllers/SeedPhraseBackupIntroductionViewController.swift
@@ -103,42 +103,3 @@ class SeedPhraseBackupIntroductionViewController: UIViewController {
         delegate?.didTapBackupWallet(inViewController: self)
     }
 }
-
-extension UIViewController {
-    func hideNavigationBarTopSeparatorLine() {
-        guard let navigationBar = navigationController?.navigationBar else { return }
-        let appearance = navigationBar.standardAppearance
-        appearance.shadowColor = .clear
-        appearance.shadowImage = nil
-        navigationBar.scrollEdgeAppearance = appearance
-        navigationBar.compactAppearance = appearance
-        navigationBar.standardAppearance = appearance
-    }
-
-    func showNavigationBarTopSeparatorLine() {
-        guard let navigationBar = navigationController?.navigationBar else { return }
-        let appearance = navigationBar.standardAppearance
-        appearance.shadowColor = Style.NavigationBar.Separator.color
-        appearance.shadowImage = nil
-        navigationBar.scrollEdgeAppearance = appearance
-        navigationBar.compactAppearance = appearance
-        navigationBar.standardAppearance = appearance
-    }
-
-    func hideNavigationBarTopSeparatorLineInScrollEdgeAppearance() {
-        guard let navigationBar = navigationController?.navigationBar else { return }
-        let appearance = navigationBar.standardAppearance
-        appearance.shadowColor = .clear
-        appearance.shadowImage = nil
-        navigationBar.scrollEdgeAppearance = appearance
-    }
-
-    func showNavigationBarTopSeparatorLineInScrollEdgeAppearance() {
-        guard let navigationBar = navigationController?.navigationBar else { return }
-        let appearance = navigationBar.standardAppearance
-        appearance.shadowColor = Style.NavigationBar.Separator.color
-        appearance.shadowImage = nil
-        navigationBar.scrollEdgeAppearance = appearance
-    }
-
-}
